fixed wrong typecasting

+++ octoprint_OctoPNP/SmdParts.py
@@ -36,7 +36,7 @@ class SmdParts():
 		return int(self._et.find("./part[@id='" + str(partnr) + "']/position").get("box"))
 
 	def getPartHeight(self, partnr):
-		return int(self._et.find("./part[@id='" + str(partnr) + "']/size").get("height"))
+		return float(self._et.find("./part[@id='" + str(partnr) + "']/size").get("height"))
 
 
 	def getPartDestination(self, partnr):
